Improved sampler.

+++ metagraph/experiments/lossless_dbg/samplers.hpp
@@ -52,8 +52,8 @@ private:
 class SubSampler : Sampler {
     SubSampler(string reference, int subsample_size, std::mt19937 generator) :
         Sampler(reference.substr(
-                std::uniform_int_distribution<>(0,reference.length()-min(subsample_size,reference.length()))(generator),
-                min(subsample_size,reference.length())
+                std::uniform_int_distribution<>(0,reference.length()-min(subsample_size,(int)reference.length()))(generator),
+                min(subsample_size,(int)reference.length())
                 )
                 generator)
                 {}
